Reorganization so WaveClientSample can compile
Mostly reorganizing and reducing dependencies so
that edu.berkeley.androidwave.waveclient is self-
contained. (And therefore can be made available to
developers of Wave client applications).  These
changes result from the realization that
Parcelables used in the IWaveServicePublic needed
to be provided in full to apps that use that
service.

Same tests pass and fail, aside from the new
WaveRecipeAuthorizationTest#testAsInfo which
fails (due to a lack of an implementation).