First code attempt at working sensing.
The entire scheduling system for sensors has now
been set up (although not supporting location
yet).  The scheduler is currently very simple,
just dropping data to avoid exceeding limits. We
are almost certain to need a better system.  This
should suffice for the initial demo.

Much of the SensorEngine is not well optimized.

This code has not yet been run, so while it now
compiles, it has not been tested yet.